requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.2.2
python-telegram-bot==22.7
openpyxl==3.1.5
pandas==3.0.2
Pillow==11.1.0
httpx==0.28.1
SQLAlchemy==2.0.49
jinja2==3.1.5
python-multipart==0.0.20
